+++ modules/api/src/main/java/com/nisovin/shopkeepers/api/shopkeeper/offers/PriceOffer.java
@@ -15,9 +15,7 @@ public interface PriceOffer {
 	/**
 	 * Creates a new {@link PriceOffer}.
 	 * <p>
-	 * If the given item stack is an {@link UnmodifiableItemStack}, it is assumed to be immutable
-	 * and therefore not copied before it is stored by the price offer. Otherwise, it is first
-	 * copied.
+	 * The given item stack is copied before it is stored by the price offer.
 	 * 
 	 * @param item
 	 *            the item being traded, not <code>null</code> or empty