+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopkeeper/offers/SKTradeOffer.java
@@ -38,9 +38,7 @@ public class SKTradeOffer extends SKTradingRecipe implements TradeOffer {
 	/**
 	 * Creates a new {@link SKTradeOffer}.
 	 * <p>
-	 * If the given item stacks are {@link UnmodifiableItemStack}s, they are assumed to be immutable
-	 * and therefore not copied before they are stored by the trade offer. Otherwise, they are first
-	 * copied.
+	 * The given item stacks are copied before they are stored by the trade offer.
 	 * 
 	 * @param resultItem
 	 *            the result item, not empty