+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopkeeper/player/sell/SKSellingPlayerShopkeeper.java
@@ -28,6 +28,7 @@ import com.nisovin.shopkeepers.util.data.property.BasicProperty;
 import com.nisovin.shopkeepers.util.data.property.Property;
 import com.nisovin.shopkeepers.util.data.serialization.InvalidDataException;
 import com.nisovin.shopkeepers.util.inventory.InventoryUtils;
+import com.nisovin.shopkeepers.util.inventory.ItemUtils;
 import com.nisovin.shopkeepers.util.java.CollectionUtils;
 import com.nisovin.shopkeepers.util.java.Validate;
 
@@ -163,7 +164,7 @@ public class SKSellingPlayerShopkeeper
 	@Override
 	public @Nullable PriceOffer getOffer(UnmodifiableItemStack tradedItem) {
 		Validate.notNull(tradedItem, "tradedItem is null");
-		return this.getOffer(tradedItem.asItemStack());
+		return this.getOffer(ItemUtils.asItemStack(tradedItem));
 	}
 
 	@Override
@@ -183,7 +184,7 @@ public class SKSellingPlayerShopkeeper
 	@Override
 	public void removeOffer(UnmodifiableItemStack tradedItem) {
 		Validate.notNull(tradedItem, "tradedItem is null");
-		this.removeOffer(tradedItem.asItemStack());
+		this.removeOffer(ItemUtils.asItemStack(tradedItem));
 	}
 
 	@Override