+++ modules/main/src/main/java/com/nisovin/shopkeepers/util/inventory/PotionUtils.java
@@ -18,6 +18,7 @@ import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import com.nisovin.shopkeepers.api.internal.util.Unsafe;
+import com.nisovin.shopkeepers.util.annotations.ReadWrite;
 import com.nisovin.shopkeepers.util.bukkit.MinecraftEnumUtils;
 import com.nisovin.shopkeepers.util.bukkit.NamespacedKeyUtils;
 import com.nisovin.shopkeepers.util.java.CollectionUtils;
@@ -33,8 +34,9 @@ public final class PotionUtils {
 	private static final Map<String, @NonNull PotionType> POTION_TYPE_ALIASES = new HashMap<>();
 	static {
 		// Removed in Bukkit 1.20.5:
-		//POTION_TYPE_ALIASES.put("empty", PotionType.UNCRAFTABLE);
-		// TODO Renamed in Bukkit 1.20.5: https://hub.spigotmc.org/stash/projects/SPIGOT/repos/bukkit/commits/8a34e009148cc297bcc9eb5c250fc4f5b071c4a7
+		// POTION_TYPE_ALIASES.put("empty", PotionType.UNCRAFTABLE);
+		// TODO Renamed in Bukkit 1.20.5:
+		// https://hub.spigotmc.org/stash/projects/SPIGOT/repos/bukkit/commits/8a34e009148cc297bcc9eb5c250fc4f5b071c4a7
 		POTION_TYPE_ALIASES.put("leaping", PotionType.JUMP);
 		POTION_TYPE_ALIASES.put("swiftness", PotionType.SPEED);
 		POTION_TYPE_ALIASES.put("healing", PotionType.INSTANT_HEAL);
@@ -207,7 +209,7 @@ public final class PotionUtils {
 	 * @return the same item stack
 	 * @see PotionMeta#setBasePotionData(PotionData)
 	 */
-	public static ItemStack setPotionData(ItemStack itemStack, PotionData potionData) {
+	public static ItemStack setPotionData(@ReadWrite ItemStack itemStack, PotionData potionData) {
 		Validate.notNull(itemStack, "itemStack is null");
 		Validate.notNull(potionData, "potionData is null");
 		ItemMeta itemMeta = itemStack.getItemMeta();