+++ modules/main/src/main/java/com/nisovin/shopkeepers/util/inventory/SKUnmodifiableItemStack.java
@@ -7,7 +7,6 @@ import org.bukkit.configuration.serialization.DelegateDeserialization;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
-import org.bukkit.material.MaterialData;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.checker.nullness.qual.PolyNull;
 
@@ -16,22 +15,12 @@ import com.nisovin.shopkeepers.util.annotations.ReadOnly;
 
 /**
  * An unmodifiable view on an {@link ItemStack}.
- * <p>
- * This class extends {@link ItemStack} so that it can be substituted in code that expects an
- * {@link ItemStack}, as long as that code does not attempt to modify the item stack. All getters
- * delegate to the wrapped item stack.
- * <p>
- * The implementation of {@link #clone()} does not return a copy of this unmodifiable item stack,
- * but instead returns a modifiable copy of the underlying item stack.
  */
 @DelegateDeserialization(ItemStack.class) // De-/Serialized as a normal modifiable ItemStack
-public class SKUnmodifiableItemStack extends ItemStack implements UnmodifiableItemStack {
+public class SKUnmodifiableItemStack implements UnmodifiableItemStack {
 
 	/**
 	 * Creates an {@link UnmodifiableItemStack} for the given {@link ItemStack}.
-	 * <p>
-	 * If the given item stack is already an {@link UnmodifiableItemStack}, this returns the given
-	 * item stack itself.
 	 * 
 	 * @param itemStack
 	 *            the item stack, can be <code>null</code>
@@ -40,9 +29,6 @@ public class SKUnmodifiableItemStack extends ItemStack implements UnmodifiableIt
 	 */
 	public static @PolyNull UnmodifiableItemStack of(@ReadOnly @PolyNull ItemStack itemStack) {
 		if (itemStack == null) return null;
-		if (itemStack instanceof UnmodifiableItemStack) {
-			return (UnmodifiableItemStack) itemStack;
-		}
 		return new SKUnmodifiableItemStack(itemStack);
 	}
 
@@ -52,32 +38,23 @@ public class SKUnmodifiableItemStack extends ItemStack implements UnmodifiableIt
 
 	private SKUnmodifiableItemStack(@ReadOnly ItemStack itemStack) {
 		assert itemStack != null;
-		assert !(itemStack instanceof UnmodifiableItemStack);
 		this.delegate = itemStack;
 	}
 
 	/**
 	 * Gets the underlying {@link ItemStack}.
 	 * <p>
-	 * This method is only meant to be used for internal debugging or testing purposes! The caller
-	 * is expected to not modify or expose the returned item stack.
+	 * The caller is expected to not modify or expose the returned item stack.
 	 * 
 	 * @return the underlying item stack
-	 * @deprecated For internal debugging and testing purposes only!
+	 * @deprecated For internal use only, when we know that the item stack is not going to be
+	 *             modified and we can therefore avoid copying it before using it in a context that
+	 *             requires an {@link ItemStack}.
 	 */
 	@Deprecated
 	public ItemStack getInternalItemStack() {
 		return delegate;
 	}
-
-	private UnsupportedOperationException unmodifiableException() {
-		return new UnsupportedOperationException("This ItemStack cannot be modified!");
-	}
-
-	// Note: Any methods that the base class implements by delegating to other methods do not
-	// necessarily need to be overridden here. However, we delegate all read-only methods to the
-	// underlying item stack nevertheless, because the underlying item stack might provide an
-	// improved implementation specific to its concrete type of item stack (e.g. CraftItemStack).
 
 	@Override
 	public ItemStack copy() {
@@ -90,18 +67,8 @@ public class SKUnmodifiableItemStack extends ItemStack implements UnmodifiableIt
 	}
 
 	@Override
-	public ItemStack asItemStack() {
-		return this;
-	}
-
-	@Override
 	public Material getType() {
 		return delegate.getType();
-	}
-
-	@Override
-	public void setType(Material type) {
-		throw unmodifiableException();
 	}
 
 	@Override
@@ -110,97 +77,46 @@ public class SKUnmodifiableItemStack extends ItemStack implements UnmodifiableIt
 	}
 
 	@Override
-	public void setAmount(int amount) {
-		throw unmodifiableException();
-	}
-
-	@Override
-	public @Nullable MaterialData getData() {
-		return delegate.getData();
-	}
-
-	@Override
-	public void setData(@Nullable MaterialData data) {
-		throw unmodifiableException();
-	}
-
-	@Override
-	public void setDurability(short durability) {
-		throw unmodifiableException();
-	}
-
-	@Override
-	public short getDurability() {
-		return delegate.getDurability();
-	}
-
-	@Override
 	public int getMaxStackSize() {
 		return delegate.getMaxStackSize();
 	}
 
-	// GH #842: Added in Paper 1.20.1 build 185 (deb92c212951ea5051fe18085f0e11ba70bf2f4b).
-	// We need to override this, because the default implementation references the type and amount
-	// fields directly, instead of via the getters.
-	// @Override // Not possible (we don't depend on Paper), but also not necessary for the
-	// override to have an effect.
-	public boolean isEmpty() {
-		return getType().isAir() || getAmount() <= 0;
-	}
-
-	// The given item stack might also be another unmodifiable item stack.
 	@Override
 	public boolean isSimilar(@ReadOnly @Nullable ItemStack itemStack) {
-		if (itemStack == this) return true;
-		if (itemStack instanceof UnmodifiableItemStack) {
-			// If this unmodifiable item stack's delegate is a CraftItemStack and the given item
-			// stack is an UnmodifiableItemStack, invoking CraftItemStack#isSimilar(itemStack) may
-			// return a wrong result. Instead, we therefore want to compare the underlying delegate
-			// item stacks with each other. Unlike when comparing CraftItemStacks with unmodifiable
-			// item stacks, the order of this comparison does not matter.
-			// This is expected to not modify or expose the passed item stack:
-			return itemStack.isSimilar(delegate);
-		} else {
-			return delegate.isSimilar(itemStack);
-		}
+		return delegate.isSimilar(itemStack);
 	}
 
 	@Override
 	public boolean isSimilar(@Nullable UnmodifiableItemStack itemStack) {
 		if (itemStack == null) return false;
 		if (itemStack == this) return true;
-		// Compare the underlying delegate item stacks:
+		// Compare the underlying item stacks:
 		// The order in which these delegate item stacks are compared does not matter.
 		// This is expected to not modify or expose the passed item stack:
 		return itemStack.isSimilar(delegate);
 	}
 
-	// TODO This method cannot be correctly implemented for CraftItemStack, because CraftItemStack
-	// always returns false when being compared to an unknown type of ItemStack (i.e. this method
-	// cannot fulfill the symmetry requirement of equals).
+	@Override
+	public boolean equals(@ReadOnly @Nullable ItemStack itemStack) {
+		return delegate.equals(itemStack);
+	}
+
 	@Override
 	public boolean equals(@ReadOnly @Nullable Object obj) {
 		if (this == obj) return true;
 		if (obj == null) return false;
-		if (obj instanceof UnmodifiableItemStack) {
-			UnmodifiableItemStack other = (UnmodifiableItemStack) obj;
-			// Compare the underlying delegate item stacks:
-			// The order in which these delegate item stacks are compared does not matter.
-			// This is expected to not modify or expose the passed item stack:
-			return other.equals(delegate);
-		} else {
-			return delegate.equals(obj);
-		}
+		if (!(obj instanceof UnmodifiableItemStack)) return false;
+
+		UnmodifiableItemStack other = (UnmodifiableItemStack) obj;
+		// Compare the underlying item stacks:
+		// The order in which these delegate item stacks are compared does not matter.
+		// This is expected to not modify or expose the passed item stack:
+		return other.equals(delegate);
 	}
 
 	@Override
 	public int hashCode() {
 		return delegate.hashCode();
-	}
-
-	@Override
-	public ItemStack clone() {
-		return delegate.clone();
 	}
 
 	@Override
@@ -218,20 +134,6 @@ public class SKUnmodifiableItemStack extends ItemStack implements UnmodifiableIt
 		return delegate.getEnchantments();
 	}
 
-	// addEnchantments: Delegates to other methods.
-	// addEnchantment: Delegates to other methods.
-	// addUnsafeEnchantments: Delegates to other methods.
-
-	@Override
-	public void addUnsafeEnchantment(Enchantment enchantment, int level) {
-		throw unmodifiableException();
-	}
-
-	@Override
-	public int removeEnchantment(Enchantment enchantment) {
-		throw unmodifiableException();
-	}
-
 	@Override
 	public Map<String, Object> serialize() {
 		return delegate.serialize();
@@ -246,17 +148,6 @@ public class SKUnmodifiableItemStack extends ItemStack implements UnmodifiableIt
 	public boolean hasItemMeta() {
 		return delegate.hasItemMeta();
 	}
-
-	@Override
-	public boolean setItemMeta(@ReadOnly @Nullable ItemMeta itemMeta) {
-		throw unmodifiableException();
-	}
-
-	// TODO Added in Spigot 1.20.4/5
-	/*@Override
-	public boolean removeEnchantments() {
-		throw unmodifiableException();
-	}*/
 
 	@Override
 	public String toString() {